requirements.txt
# Make sure clams-python version is explicitly specified, at least the lower bound
clams-python==1.2.6
clams-utils
nltk==3.8.1
gensim==4.3.2
scipy==1.12